+++ pypesto/objective/petab_import.py
@@ -104,9 +104,12 @@ class PetabImporter:
                 f"Refusing to remove {self.output_folder} for model "
                 f"compilation: Not a folder.")
 
+        # add module to path
+        if self.output_folder not in sys.path:
+            sys.path.insert(0, self.output_folder)
+
         # compile
-        if force_compile or not os.path.exists(self.output_folder) or \
-                not os.listdir(self.output_folder):
+        if self._must_compile(force_compile):
             logger.info(f"Compiling amici model to folder "
                         f"{self.output_folder}.")
             self.compile_model()
@@ -114,10 +117,13 @@ class PetabImporter:
             logger.info(f"Using existing amici model in folder "
                         f"{self.output_folder}.")
 
-        # add module to path
-        if self.output_folder not in sys.path:
-            sys.path.insert(0, self.output_folder)
+        return self._create_model()
 
+    def _create_model(self):
+        """
+        No checks, no compilation, just load the model module and return
+        the model.
+        """
         # load moduĺe
         model_module = importlib.import_module(self.model_name)
 
@@ -125,6 +131,29 @@ class PetabImporter:
         model = model_module.getModel()
 
         return model
+
+    def _must_compile(self, force_compile: bool):
+        """
+        Check whether the model needs to be compiled first.
+        """
+        # asked by user
+        if force_compile:
+            return True
+
+        # folder does not exist
+        if not os.path.exists(self.output_folder) or \
+                not os.listdir(self.output_folder):
+            return True
+
+        # try to import (in particular checks version)
+        try:
+            # importing will already raise an exception if version wrong
+            importlib.import_module(self.model_name)
+        except RuntimeError:
+            return True
+
+        # no need to (re-)compile
+        return False
 
     def compile_model(self):
         """
@@ -598,8 +627,7 @@ class PetabAmiciObjective(AmiciObjective):
     def __getstate__(self):
         state = {}
         for key in set(self.__dict__.keys()) - \
-                set(['amici_model', 'amici_solver', 'edatas',
-                     'preequilibration_edatas']):
+                set(['amici_model', 'amici_solver', 'edatas']):
             state[key] = self.__dict__[key]
         return state
 
@@ -615,26 +643,15 @@ class PetabAmiciObjective(AmiciObjective):
         self.amici_solver = solver
         self.edatas = edatas
 
-        if self.preprocess_edatas:
-            self.init_preequilibration_edatas(edatas)
-        else:
-            self.preequilibration_edatas = None
-
     def __deepcopy__(self, memodict=None):
         other = self.__class__.__new__(self.__class__)
 
         for key in set(self.__dict__.keys()) - \
-                set(['amici_model', 'amici_solver', 'edatas',
-                     'preequilibration_edatas']):
+                set(['amici_model', 'amici_solver', 'edatas']):
             other.__dict__[key] = copy.deepcopy(self.__dict__[key])
 
         other.amici_model = amici.ModelPtr(self.amici_model.clone())
         other.amici_solver = amici.SolverPtr(self.amici_solver.clone())
         other.edatas = [amici.ExpData(data) for data in self.edatas]
-
-        if self.preprocess_edatas:
-            other.init_preequilibration_edatas(other.edatas)
-        else:
-            other.preequilibration_edatas = None
 
         return other