+++ pypesto/visualize/waterfall.py
@@ -41,7 +41,7 @@ def waterfall(results,
         May be logarithmic or linear ('log10' or 'lin')
 
     offset_y:
-        offset for the y-axis, if this is supposed to be in log10-scale
+        offset for the y-axis, if it is supposed to be in log10-scale
 
     start_indices: list or int
         list of integers specifying the multistart to be plotted or
@@ -75,24 +75,25 @@ def waterfall(results,
     # loop over results
     for j, result in enumerate(results):
         # extract specific cost function values from result
-        fvals = get_fvals(result, scale_y, offset_y, start_indices)
+        (fvals, offset_y) = get_fvals(result, scale_y, offset_y, start_indices)
         max_len_fvals = np.max([max_len_fvals, len(fvals)])
 
         # call lowlevel plot routine
-        ax = waterfall_lowlevel(fvals=fvals, scale_y=scale_y, ax=ax, size=size,
+        ax = waterfall_lowlevel(fvals=fvals, scale_y=scale_y,
+                                offset_y=offset_y, ax=ax, size=size,
                                 colors=colors[j], legend_text=legends[j])
 
     # parse and apply plotting options
     ref = create_references(references=reference)
 
     # apply changes specified be the user to the axis object
-    ax = handle_options(ax, max_len_fvals, ref, y_limits)
+    ax = handle_options(ax, max_len_fvals, ref, y_limits, offset_y)
 
     return ax
 
 
-def waterfall_lowlevel(fvals, scale_y='log10', ax=None, size=(18.5, 10.5),
-                       colors=None, legend_text=None):
+def waterfall_lowlevel(fvals, scale_y='log10', offset_y=0., ax=None,
+                       size=(18.5, 10.5), colors=None, legend_text=None):
     """
     Plot waterfall plot using list of function values.
 
@@ -104,6 +105,9 @@ def waterfall_lowlevel(fvals, scale_y='log10', ax=None, size=(18.5, 10.5),
 
     scale_y: str, optional
         May be logarithmic or linear ('log10' or 'lin')
+
+    offset_y:
+        offset for the y-axis, if it is supposed to be in log10-scale
 
     ax: matplotlib.Axes, optional
         Axes object to use.
@@ -175,7 +179,10 @@ def waterfall_lowlevel(fvals, scale_y='log10', ax=None, size=(18.5, 10.5),
 
     # labels
     ax.set_xlabel('Ordered optimizer run')
-    ax.set_ylabel('Function value')
+    if offset_y == .0:
+        ax.set_ylabel('Function value')
+    else:
+        ax.set_ylabel('Offsetted function value (relative to best start)')
     ax.set_title('Waterfall plot')
     if legend_text is not None:
         ax.legend()
@@ -197,7 +204,7 @@ def get_fvals(result, scale_y, offset_y, start_indices):
         May be logarithmic or linear ('log10' or 'lin')
 
     offset_y:
-        offset for the y-axis, if this is supposed to be in log10-scale
+        offset for the y-axis, if it is supposed to be in log10-scale
 
     start_indices: list or int
         list of integers specifying the multistart to be plotted or
@@ -208,6 +215,9 @@ def get_fvals(result, scale_y, offset_y, start_indices):
 
     fvals: ndarray
         function values
+
+    offset_y:
+        offset for the y-axis, if this is supposed to be in log10-scale
     """
 
     # extract cost function values from result
@@ -240,10 +250,10 @@ def get_fvals(result, scale_y, offset_y, start_indices):
         fvals += offset_y * np.ones(fvals.shape)
 
     # get only the indices which the user asked for
-    return fvals
+    return fvals, offset_y
 
 
-def handle_options(ax, max_len_fvals, ref, y_limits):
+def handle_options(ax, max_len_fvals, ref, y_limits, offset_y):
     """
     Get the limits for the y-axis, plots the reference points, will do
     more at a later time point. This function is there to apply whatever
@@ -265,6 +275,9 @@ def handle_options(ax, max_len_fvals, ref, y_limits):
     y_limits: float or ndarray, optional
         maximum value to be plotted on the y-axis, or y-limits
 
+    offset_y:
+        offset for the y-axis, if it is supposed to be in log10-scale
+
     Returns
     -------
 
@@ -277,7 +290,13 @@ def handle_options(ax, max_len_fvals, ref, y_limits):
 
     # handle reference points
     for i_ref in ref:
-        ax.plot([0, max_len_fvals - 1], [i_ref.fval, i_ref.fval], '--',
+        # plot reference point as line
+        ax.plot([0, max_len_fvals - 1],
+                [i_ref.fval + offset_y, i_ref.fval + offset_y], '--',
                 color=i_ref.color, label=i_ref.legend)
+
+        # create legend for reference points
+        if i_ref.legend is not None:
+            ax.legend()
 
     return ax