+++ test/test_sbml_conversion.py
@@ -18,6 +18,9 @@ optimizers = {
     'dlib': ['default']
 }
 
+ATOL = 1e-2
+RTOL = 1e-3
+
 
 class AmiciObjectiveTest(unittest.TestCase):
     def runTest(self):
@@ -26,24 +29,24 @@ class AmiciObjectiveTest(unittest.TestCase):
             x0 = list(model.getParameters())
             df = objective.check_grad(
                 x0,
-                eps=1e-5,
+                eps=1e-3,
                 verbosity=0,
                 mode=pypesto.objective.constants.MODE_FUN
             )
             print("relative errors MODE_FUN: ", df.rel_err.values)
             print("absolute errors MODE_FUN: ", df.abs_err.values)
-            self.assertTrue(np.all(df.rel_err.values < 1e-2))
-            self.assertTrue(np.all(df.abs_err.values < 1e-1))
+            self.assertTrue(np.all((df.rel_err.values < RTOL) |
+                                   (df.abs_err.values < ATOL)))
             df = objective.check_grad(
                 x0,
-                eps=1e-5,
+                eps=1e-3,
                 verbosity=0,
                 mode=pypesto.objective.constants.MODE_RES
             )
             print("relative errors MODE_RES: ", df.rel_err.values)
             print("absolute errors MODE_RES: ", df.rel_err.values)
-            self.assertTrue(np.all(df.rel_err.values < 1e-2))
-            self.assertTrue(np.all(df.abs_err.values < 1e-2))
+            self.assertTrue(np.all((df.rel_err.values < RTOL) |
+                                   (df.abs_err.values < ATOL)))
 
             for library in optimizers.keys():
                 for method in optimizers[library]:
@@ -58,7 +61,7 @@ class AmiciObjectiveTest(unittest.TestCase):
                                     library,
                                     method,
                                     fp,
-                                    1)
+                                    2)
 
 
 def parameter_estimation(
@@ -89,10 +92,14 @@ def parameter_estimation(
                               x_fixed_vals=[pars[idx] for idx in fixed_pars]
                               )
 
-    optimize_options = pypesto.OptimizeOptions(allow_failed_starts=False)
+    optimize_options = pypesto.OptimizeOptions(
+        allow_failed_starts=False,
+        startpoint_resample=True,
+    )
 
     results = pypesto.minimize(
-        problem, optimizer, n_starts, options=optimize_options)
+        problem, optimizer, n_starts, options=optimize_options,
+    )
     results = results.optimize_result.list